+++ packages/cli/src/client-config/client_config_generator_adapter.ts
@@ -6,8 +6,11 @@ import {
   generateClientConfigToFile,
 } from '@aws-amplify/client-config';
 import { DeployedBackendIdentifier } from '@aws-amplify/deployed-backend-client';
-import { AwsCredentialIdentityProvider } from '@aws-sdk/types';
 import { printer } from '@aws-amplify/cli-core';
+import { AWSClientProvider } from '@aws-amplify/plugin-types';
+import { S3Client } from '@aws-sdk/client-s3';
+import { AmplifyClient } from '@aws-sdk/client-amplify';
+import { CloudFormationClient } from '@aws-sdk/client-cloudformation';
 
 /**
  * Adapts static generateClientConfigToFile from @aws-amplify/client-config call to make it injectable and testable.
@@ -17,7 +20,11 @@ export class ClientConfigGeneratorAdapter {
    * Creates new adapter for generateClientConfigToFile from @aws-amplify/client-config.
    */
   constructor(
-    private readonly awsCredentialProvider: AwsCredentialIdentityProvider
+    private readonly awsClientProvider: AWSClientProvider<{
+      getS3Client: S3Client;
+      getAmplifyClient: AmplifyClient;
+      getCloudFormationClient: CloudFormationClient;
+    }>
   ) {}
   /**
    * Generates the client configuration for a given backend
@@ -27,9 +34,9 @@ export class ClientConfigGeneratorAdapter {
     version: ClientConfigVersion
   ): Promise<ClientConfig> => {
     return generateClientConfig(
-      this.awsCredentialProvider,
       backendIdentifier,
-      version
+      version,
+      this.awsClientProvider
     );
   };
 
@@ -44,11 +51,11 @@ export class ClientConfigGeneratorAdapter {
     format?: ClientConfigFormat
   ): Promise<void> => {
     const { filesWritten } = await generateClientConfigToFile(
-      this.awsCredentialProvider,
       backendIdentifier,
       version,
       outDir,
-      format
+      format,
+      this.awsClientProvider
     );
 
     filesWritten.forEach((file) => printer.log(`File written: ${file}`));