+++ packages/cli/src/commands/generate/config/generate_config_command.test.ts
@@ -1,5 +1,4 @@
 import { beforeEach, describe, it, mock } from 'node:test';
-import { fromNodeProviderChain } from '@aws-sdk/credential-providers';
 import { GenerateConfigCommand } from './generate_config_command.js';
 import { ClientConfigFormat } from '@aws-amplify/client-config';
 import yargs, { CommandModule } from 'yargs';
@@ -9,11 +8,16 @@ import { AppBackendIdentifierResolver } from '../../../backend-identifier/backen
 import { ClientConfigGeneratorAdapter } from '../../../client-config/client_config_generator_adapter.js';
 import { BackendIdentifierResolverWithFallback } from '../../../backend-identifier/backend_identifier_with_sandbox_fallback.js';
 import { SandboxBackendIdResolver } from '../../sandbox/sandbox_id_resolver.js';
+import { S3Client } from '@aws-sdk/client-s3';
+import { AmplifyClient } from '@aws-sdk/client-amplify';
+import { CloudFormationClient } from '@aws-sdk/client-cloudformation';
 
 void describe('generate config command', () => {
-  const clientConfigGeneratorAdapter = new ClientConfigGeneratorAdapter(
-    fromNodeProviderChain()
-  );
+  const clientConfigGeneratorAdapter = new ClientConfigGeneratorAdapter({
+    getS3Client: () => new S3Client(),
+    getAmplifyClient: () => new AmplifyClient(),
+    getCloudFormationClient: () => new CloudFormationClient(),
+  });
 
   const generateClientConfigMock = mock.method(
     clientConfigGeneratorAdapter,