+++ packages/model-generator/src/create_graphql_models_generator.ts
@@ -1,23 +1,34 @@
-import { AwsCredentialIdentityProvider } from '@aws-sdk/types';
-import { S3Client } from '@aws-sdk/client-s3';
 import {
   BackendOutputClientFactory,
   DeployedBackendIdentifier,
 } from '@aws-amplify/deployed-backend-client';
+
 import { graphqlOutputKey } from '@aws-amplify/backend-output-schemas';
 import { AppsyncGraphqlGenerationResult } from './appsync_graphql_generation_result.js';
 import { StackMetadataGraphqlModelsGenerator } from './graphql_models_generator.js';
 import { GraphqlModelsGenerator } from './model_generator.js';
 import { S3StringObjectFetcher } from './s3_string_object_fetcher.js';
+import { AWSClientProvider } from '@aws-amplify/plugin-types';
+import { S3Client } from '@aws-sdk/client-s3';
+import { AmplifyClient } from '@aws-sdk/client-amplify';
+import { CloudFormationClient } from '@aws-sdk/client-cloudformation';
 
 export type GraphqlModelsGeneratorFactoryParams =
   | {
       backendIdentifier: DeployedBackendIdentifier;
-      credentialProvider: AwsCredentialIdentityProvider;
+      awsClientProvider: AWSClientProvider<{
+        getS3Client: S3Client;
+        getAmplifyClient: AmplifyClient;
+        getCloudFormationClient: CloudFormationClient;
+      }>;
     }
   | {
       modelSchemaS3Uri: string;
-      credentialProvider: AwsCredentialIdentityProvider;
+      awsClientProvider: AWSClientProvider<{
+        getS3Client: S3Client;
+        getAmplifyClient: AmplifyClient;
+        getCloudFormationClient: CloudFormationClient;
+      }>;
     };
 
 /**
@@ -34,7 +45,11 @@ export const createGraphqlModelsGenerator = (
 
 export type GraphqlModelsFromBackendIdentifierParams = {
   backendIdentifier: DeployedBackendIdentifier;
-  credentialProvider: AwsCredentialIdentityProvider;
+  awsClientProvider: AWSClientProvider<{
+    getS3Client: S3Client;
+    getAmplifyClient: AmplifyClient;
+    getCloudFormationClient: CloudFormationClient;
+  }>;
 };
 
 /**
@@ -42,24 +57,28 @@ export type GraphqlModelsFromBackendIdentifierParams = {
  */
 const createGraphqlModelsGeneratorFromBackendIdentifier = ({
   backendIdentifier,
-  credentialProvider,
+  awsClientProvider,
 }: GraphqlModelsFromBackendIdentifierParams): GraphqlModelsGenerator => {
   if (!backendIdentifier) {
     throw new Error('`backendIdentifier` must be defined');
   }
-  if (!credentialProvider) {
-    throw new Error('`credentialProvider` must be defined');
+  if (!awsClientProvider) {
+    throw new Error('`awsClientProvider` must be defined');
   }
 
   return new StackMetadataGraphqlModelsGenerator(
-    () => getModelSchema(backendIdentifier, credentialProvider),
+    () => getModelSchema(backendIdentifier, awsClientProvider),
     (fileMap) => new AppsyncGraphqlGenerationResult(fileMap)
   );
 };
 
 export type GraphqlModelsFromS3UriGeneratorFactoryParams = {
   modelSchemaS3Uri: string;
-  credentialProvider: AwsCredentialIdentityProvider;
+  awsClientProvider: AWSClientProvider<{
+    getS3Client: S3Client;
+    getAmplifyClient: AmplifyClient;
+    getCloudFormationClient: CloudFormationClient;
+  }>;
 };
 
 /**
@@ -67,28 +86,30 @@ export type GraphqlModelsFromS3UriGeneratorFactoryParams = {
  */
 export const createGraphqlModelsFromS3UriGenerator = ({
   modelSchemaS3Uri,
-  credentialProvider,
+  awsClientProvider,
 }: GraphqlModelsFromS3UriGeneratorFactoryParams): GraphqlModelsGenerator => {
   if (!modelSchemaS3Uri) {
     throw new Error('`modelSchemaS3Uri` must be defined');
   }
-  if (!credentialProvider) {
-    throw new Error('`credentialProvider` must be defined');
+  if (!awsClientProvider) {
+    throw new Error('`awsClientProvider` must be defined');
   }
-
   return new StackMetadataGraphqlModelsGenerator(
-    () => getModelSchemaFromS3Uri(modelSchemaS3Uri, credentialProvider),
+    () => getModelSchemaFromS3Uri(modelSchemaS3Uri, awsClientProvider),
     (fileMap) => new AppsyncGraphqlGenerationResult(fileMap)
   );
 };
 
 const getModelSchema = async (
   backendIdentifier: DeployedBackendIdentifier,
-  credentialProvider: AwsCredentialIdentityProvider
+  awsClientProvider: AWSClientProvider<{
+    getS3Client: S3Client;
+    getAmplifyClient: AmplifyClient;
+    getCloudFormationClient: CloudFormationClient;
+  }>
 ): Promise<string> => {
-  const backendOutputClient = BackendOutputClientFactory.getInstance({
-    credentials: credentialProvider,
-  });
+  const backendOutputClient =
+    BackendOutputClientFactory.getInstance(awsClientProvider);
   const output = await backendOutputClient.getOutput(backendIdentifier);
   const modelSchemaS3Uri =
     output[graphqlOutputKey]?.payload.amplifyApiModelSchemaS3Uri;
@@ -96,16 +117,17 @@ const getModelSchema = async (
     throw new Error(`Cannot find model schema at amplifyApiModelSchemaS3Uri`);
   }
 
-  return await getModelSchemaFromS3Uri(modelSchemaS3Uri, credentialProvider);
+  return await getModelSchemaFromS3Uri(modelSchemaS3Uri, awsClientProvider);
 };
 
 const getModelSchemaFromS3Uri = async (
   modelSchemaS3Uri: string,
-  credentialProvider: AwsCredentialIdentityProvider
+  awsClientProvider: AWSClientProvider<{
+    getS3Client: S3Client;
+  }>
 ): Promise<string> => {
-  const s3Client = new S3Client({
-    credentials: credentialProvider,
-  });
-  const schemaFetcher = new S3StringObjectFetcher(s3Client);
+  const schemaFetcher = new S3StringObjectFetcher(
+    awsClientProvider.getS3Client()
+  );
   return await schemaFetcher.fetch(modelSchemaS3Uri);
 };