+++ packages/backend-auth/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aws-amplify/backend-auth
 
+## 0.5.0-beta.13
+
+### Patch Changes
+
+- ef111b4: Add friendly-name tag to resources
+- Updated dependencies [db23a3f]
+  - @aws-amplify/plugin-types@0.9.0-beta.3
+  - @aws-amplify/backend-output-storage@0.4.0-beta.8
+  - @aws-amplify/auth-construct-alpha@0.6.0-beta.13
+
 ## 0.5.0-beta.12
 
 ### Patch Changes