+++ packages/deployed-backend-client/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @aws-amplify/deployed-backend-client
 
+## 0.4.0-beta.10
+
+### Patch Changes
+
+- Updated dependencies [ef111b4]
+  - @aws-amplify/platform-core@0.5.0-beta.7
+
+## 0.4.0-beta.9
+
+### Minor Changes
+
+- fe46848: Allow passing clients in to client-config from browser context.
+
 ## 0.4.0-beta.8
 
 ### Patch Changes