+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @aws-amplify/sandbox
 
+## 0.5.2-beta.19
+
+### Patch Changes
+
+- Updated dependencies [ef111b4]
+  - @aws-amplify/platform-core@0.5.0-beta.7
+  - @aws-amplify/backend-deployer@0.5.1-beta.8
+  - @aws-amplify/backend-secret@0.4.5-beta.7
+  - @aws-amplify/cli-core@0.5.0-beta.12
+  - @aws-amplify/client-config@0.9.0-beta.15
+  - @aws-amplify/deployed-backend-client@0.4.0-beta.10
+
+## 0.5.2-beta.18
+
+### Patch Changes
+
+- e90f066: support single sandbox deployment with --once flag
+- Updated dependencies [fe46848]
+- Updated dependencies [4f66069]
+  - @aws-amplify/deployed-backend-client@0.4.0-beta.9
+  - @aws-amplify/client-config@0.9.0-beta.14
+
 ## 0.5.2-beta.17
 
 ### Patch Changes